+++ src/svd/interface/svdbasic.c
@@ -65,14 +65,14 @@ PetscErrorCode SVDCreate(MPI_Comm comm,SVD *outsvd)
   svd->nini             = 0;
   svd->ninil            = 0;
   svd->tol              = PETSC_DEFAULT;
-  svd->conv             = SVD_CONV_REL;
+  svd->conv             = (SVDConv)-1;
   svd->stop             = SVD_STOP_BASIC;
   svd->which            = SVD_LARGEST;
   svd->problem_type     = (SVDProblemType)0;
   svd->impltrans        = PETSC_FALSE;
   svd->trackall         = PETSC_FALSE;
 
-  svd->converged        = SVDConvergedRelative;
+  svd->converged        = NULL;
   svd->convergeduser    = NULL;
   svd->convergeddestroy = NULL;
   svd->stopping         = SVDStoppingBasic;