+++ src/svd/interface/svdview.c
@@ -275,6 +275,13 @@ static PetscErrorCode SVDErrorView_DETAIL(SVD svd,SVDErrorType etype,PetscViewer
     case SVD_ERROR_RELATIVE:
       ierr = PetscSNPrintf(ex,sizeof(ex)," relative error");CHKERRQ(ierr);
       break;
+    case SVD_ERROR_NORM:
+      if (svd->isgeneralized) {
+        ierr = PetscSNPrintf(ex,sizeof(ex)," ||r||/||[A;B]||");CHKERRQ(ierr);
+      } else {
+        ierr = PetscSNPrintf(ex,sizeof(ex),"  ||r||/||A||");CHKERRQ(ierr);
+      }
+      break;
   }
   ierr = PetscViewerASCIIPrintf(viewer,"%s          sigma           %s\n%s",sep,ex,sep);CHKERRQ(ierr);
   for (i=0;i<svd->nconv;i++) {
@@ -317,7 +324,8 @@ static PetscErrorCode SVDErrorView_MATLAB(SVD svd,SVDErrorType etype,PetscViewer
 
    Options Database Key:
 +  -svd_error_absolute - print absolute errors of each singular triplet
--  -svd_error_relative - print relative errors of each singular triplet
+.  -svd_error_relative - print relative errors of each singular triplet
+-  -svd_error_norm     - print errors relative to the matrix norms of each singular triplet
 
    Notes:
    By default, this function checks the error of all singular triplets and prints
@@ -399,6 +407,13 @@ PetscErrorCode SVDErrorViewFromOptions(SVD svd)
   if (flg) {
     ierr = PetscViewerPushFormat(viewer,format);CHKERRQ(ierr);
     ierr = SVDErrorView(svd,SVD_ERROR_RELATIVE,viewer);CHKERRQ(ierr);
+    ierr = PetscViewerPopFormat(viewer);CHKERRQ(ierr);
+    ierr = PetscViewerDestroy(&viewer);CHKERRQ(ierr);
+  }
+  ierr = PetscOptionsGetViewer(PetscObjectComm((PetscObject)svd),((PetscObject)svd)->options,((PetscObject)svd)->prefix,"-svd_error_norm",&viewer,&format,&flg);CHKERRQ(ierr);
+  if (flg) {
+    ierr = PetscViewerPushFormat(viewer,format);CHKERRQ(ierr);
+    ierr = SVDErrorView(svd,SVD_ERROR_NORM,viewer);CHKERRQ(ierr);
     ierr = PetscViewerPopFormat(viewer);CHKERRQ(ierr);
     ierr = PetscViewerDestroy(&viewer);CHKERRQ(ierr);
   }