+++ src/svd/tutorials/ex45.c
@@ -134,10 +134,10 @@ int main(int argc,char **argv)
     ierr = VecDuplicateVecs(v,nconv,&V);CHKERRQ(ierr);
 
     /*
-       Display singular values and relative errors
+       Display singular values and errors relative to the norms
     */
     ierr = PetscPrintf(PETSC_COMM_WORLD,
-         "          sigma           relative error\n"
+         "          sigma           ||r||/||[A;B]||\n"
          "  --------------------- ------------------\n");CHKERRQ(ierr);
     for (i=0;i<nconv;i++) {
       /*
@@ -152,7 +152,7 @@ int main(int argc,char **argv)
       /*
          Compute the error associated to each singular triplet
       */
-      ierr = SVDComputeError(svd,i,SVD_ERROR_RELATIVE,&error);CHKERRQ(ierr);
+      ierr = SVDComputeError(svd,i,SVD_ERROR_NORM,&error);CHKERRQ(ierr);
 
       ierr = PetscPrintf(PETSC_COMM_WORLD,"       % 6f      ",(double)sigma);CHKERRQ(ierr);
       ierr = PetscPrintf(PETSC_COMM_WORLD,"   % 12g\n",(double)error);CHKERRQ(ierr);